feat: add tests for input/output files handling
