+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "nodejs2021q4",
   "version": "1.0.0",
-  "description": "TASK 1",
+  "description": "NODEJS2021",
   "main": "ExistingFileError.js",
   "scripts": {
     "test": "jest",
