+++ utils/index.js
@@ -1,10 +1,8 @@
 const { showError } = require('./showError');
 const { getChainConfig } = require('./getChainConfig');
-const {
-  getArrayOfTransformStreams,
-  getInputFile,
-  getOutputFile,
-} = require('./utils');
+const { getInputFile } = require('./getInputFile');
+const { getOutputFile } = require('./getOutputFile');
+const { getArrayOfTransformStreams } = require('./getArrayOfTransformStreams');
 
 module.exports = {
   getArrayOfTransformStreams,
